requirements.txt
torch
pillow
numpy
requests